+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Changed
 - update check-graylog2-alive.rb to accept an `--apipath` argument to specify the path of the transport api
 - update -alive lifecycle check because there are now multiple valid lifecyle states
-- update sensu-plugin dep to '~> 1.2'
+- drop ruby 1.9.3 support
+- update sensu-plugin dep to '1.4.4'
 - add some tests, this pulled in webmock to mock restclient calls
 - update readme