+++ CHANGELOG.md
@@ -17,6 +17,14 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
   - add version support for pre/post 2.1.0 buffer metrics
   - add apipath support
   - add tests
+- metics-graylog
+  - port python to ruby
+  - py file is now a binstub for rb file
+  - add --all flag for more stats (still some work to do here)
+  - add apipath support
+  - add tests
+
+
 
 ## [0.1.0] - 2016-01-29
 ### Added